replace int to size_t on wl_def.h

+++ wl_def.h
@@ -962,7 +962,7 @@ typedef struct objstruct
 
     unsigned int    flags;              // FL_SHOOTABLE, etc
 
-    int     distance;           // if negative, wait for that door to open
+    size_t     distance;           // if negative, wait for that door to open
     dirtype     dir;
 
     fixed       x, y;
